ed_quant_engine/requirements.txt
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.4
pandas_ta==0.3.14b0
scikit-learn==1.4.0
nltk==3.8.1
python-telegram-bot==20.8
schedule==1.2.1
matplotlib==3.8.2
seaborn==0.13.2
requests==2.31.0
joblib==1.3.2
feedparser==6.0.11
python-dotenv==1.0.1
